LocListener for finding current location
Added the beginning of a query to take the current location and return radial search of universities in that region.

Created a listener to the geofire listener ran the dependancies and the commits added google play servicees and then created a screen to add the button that will find the location and return the universities.